requirements.txt
# CareBox Digital Twin - Requirements
# Python 3.11+

# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# Environment & Config
python-dotenv==1.0.0

# Google Sheets API
gspread==5.12.0
google-auth==2.23.4

# QR Code Generation
qrcode==7.4.2
Pillow==10.1.0

# Security
bleach==6.1.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
